requirements.txt
boto3==1.20.42
fuzzywuzzy==0.18.0
librosa==0.8.0
matplotlib==3.3.3
moviepy==1.0.3
names==0.3.0
numpy==1.21.5
opencv_python==4.5.5.62
pandas==1.3.5
panns_inference==0.0.7
Pillow==9.0.1
pydub==0.25.1
requests==2.27.1
scipy==1.7.3
streamlit==1.5.1
torch==1.10.2
